refactor(uikit): split MiniCart to smaller files
