+++ packages/falcon-ui-kit/src/MiniCart/MiniCart.tsx
@@ -1,51 +1,12 @@
 import React from 'react';
 import { Link as RouterLink } from 'react-router-dom';
-import {
-  Icon,
-  List,
-  ListItem,
-  Box,
-  H3,
-  H4,
-  DefaultThemeProps,
-  Image,
-  Link,
-  Text,
-  Divider,
-  Button,
-  NumberInput,
-  FlexLayout
-} from '@deity/falcon-ui';
+import { Icon, List, ListItem, Box, H3, DefaultThemeProps, Text, Divider, Button, FlexLayout } from '@deity/falcon-ui';
 import { LocaleProvider } from '@deity/falcon-front-kit';
-import { I18n, T } from '@deity/falcon-i18n';
-import { RemoveCartItemMutation, UpdateCartItemMutation } from '@deity/falcon-shop-data';
+import { T } from '@deity/falcon-i18n';
 import { CloseSidebarMutation } from '../Sidebar';
 import { toGridTemplate, prettyScrollbars } from '../helpers';
-import { Price } from '../Price';
 import { MiniCartData } from './MiniCartQuery';
-
-export const MiniCartProductArea = {
-  empty: '.',
-  thumb: 'thumb',
-  price: 'price',
-  productName: 'productName',
-  modify: 'modify',
-  remove: 'remove'
-};
-
-const miniCartProductTheme: DefaultThemeProps = {
-  miniCartProduct: {
-    display: 'grid',
-    gridGap: 'xs',
-    // prettier-ignore
-    gridTemplate: toGridTemplate([
-      ['30px',                     '1fr',                     '2fr'                                    ],
-      [MiniCartProductArea.empty,  MiniCartProductArea.thumb, MiniCartProductArea.productName          ],
-      [MiniCartProductArea.remove, MiniCartProductArea.thumb, MiniCartProductArea.price,          '1fr'],
-      [MiniCartProductArea.empty,  MiniCartProductArea.thumb, MiniCartProductArea.modify               ]
-    ])
-  }
-};
+import { MiniCartProduct } from './MiniCartProduct';
 
 export const MiniCartLayoutArea = {
   title: 'title',
@@ -53,8 +14,8 @@ export const MiniCartLayoutArea = {
   cta: 'cta'
 };
 
-const miniCartLayout: DefaultThemeProps = {
-  miniCartLayout: {
+const miniCartTheme: DefaultThemeProps = {
+  miniCart: {
     display: 'grid',
     gridRowGap: 'md',
 
@@ -72,67 +33,6 @@ const miniCartLayout: DefaultThemeProps = {
   }
 };
 
-const MiniCartProduct: React.SFC<any> = ({ product }) => (
-  <Box defaultTheme={miniCartProductTheme}>
-    <Image gridArea={MiniCartProductArea.thumb} src={product.thumbnailUrl} />
-    <H4 gridArea={MiniCartProductArea.productName}>{product.name}</H4>
-    <Price gridArea={MiniCartProductArea.price} value={product.price} fontSize="md" fontWeight="bold" />
-    <RemoveCartItemMutation>
-      {removeCartItem => (
-        <Link
-          gridArea={MiniCartProductArea.remove}
-          display="flex"
-          alignItems="center"
-          onClick={() =>
-            removeCartItem({
-              variables: { input: { itemId: product.itemId } },
-              optimisticResponse: { removeCartItem: { __typename: 'RemoveCartItemPayload', itemId: product.itemId } }
-            })
-          }
-        >
-          <Icon
-            size="lg"
-            stroke="secondaryDark"
-            src="remove"
-            mr="xs"
-            css={({ theme }) => ({ ':hover': { stroke: theme.colors.primary } })}
-          />
-        </Link>
-      )}
-    </RemoveCartItemMutation>
-    <UpdateCartItemMutation>
-      {(updateCartItem, { loading, error }) => (
-        <Box gridArea={MiniCartProductArea.modify}>
-          <I18n>
-            {t => (
-              <NumberInput
-                disabled={loading}
-                min="1"
-                name="qty"
-                defaultValue={String(product.qty)}
-                aria-label={t('product.quantity')}
-                onChange={ev =>
-                  updateCartItem({
-                    variables: {
-                      input: {
-                        itemId: product.itemId,
-                        sku: product.sku,
-                        qty: parseInt(ev.target.value, 10)
-                      }
-                    }
-                  })
-                }
-              />
-            )}
-          </I18n>
-
-          {!!error && <Text color="error">{error.message}</Text>}
-        </Box>
-      )}
-    </UpdateCartItemMutation>
-  </Box>
-);
-
 const MiniCartProducts: React.SFC<any> = ({ products }) => (
   <List>
     {products.map((product: any, index: number) => (
@@ -148,7 +48,7 @@ export const MiniCart: React.SFC<MiniCartData> = ({ cart: { quoteCurrency, items
   <CloseSidebarMutation>
     {closeSidebar => (
       <LocaleProvider currency={quoteCurrency}>
-        <Box defaultTheme={miniCartLayout}>
+        <Box defaultTheme={miniCartTheme}>
           <H3 gridArea={MiniCartLayoutArea.title}>
             <T id="miniCart.title" />
           </H3>
