+++ packages/falcon-ui-kit/src/MiniCart/index.ts
@@ -1,3 +1,4 @@
 export * from './MiniCart';
 export * from './MiniCartQuery';
 export * from './MiniCartIcon';
+export * from './MiniCartProduct';
